refactor reporting schedule run router
